+++ src/activities/home/HomeActivity.h
@@ -71,6 +71,7 @@ class HomeActivity final : public Activity {
   void onFileBrowserOpen();
   void onContinueReading();
   void onRecentsOpen();
+  void onTbrListOpen();
   void onSettingsOpen();
   void onFileTransferOpen();
   void onOpdsBrowserOpen();